requirements.txt
ansible==9.4.0
ansible-core==2.16.5
ansible-vault==2.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pycparser==2.22
PyYAML==6.0.1
requests==2.31.0
resolvelib==1.0.1
smmap==5.0.1
urllib3==2.2.1
